requirements.txt
authy==2.2.3
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-phonenumber-field==2.1.0
django-tastypie==0.14.2
idna==2.8
mypy==0.650
mypy-extensions==0.4.1
phonenumbers==8.9.16
Pillow==5.3.0
python-dateutil==2.7.5
python-dotenv==0.10.0
python-mimeparse==1.6.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
simplejson==3.16.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
